Несколько ухудшилась читаемость, однако получились такие результаты
```
name     old time/op    new time/op    delta
Build-8    2.06µs ± 1%    1.88µs ± 0%   -9.05%  (p=0.008 n=5+5)

name     old alloc/op   new alloc/op   delta
Build-8    1.78kB ± 0%    1.44kB ± 0%  -18.92%  (p=0.008 n=5+5)

name     old allocs/op  new allocs/op  delta
Build-8      15.0 ± 0%      11.0 ± 0%  -26.67%  (p=0.008 n=5+5)
```

Также добавлены методы для получения размера подписи.